Change Strutil::memformat to uint64_t byte counts

+++ src/include/OpenImageIO/strutil.h
@@ -260,7 +260,7 @@ std::string OIIO_API vformat (const char *fmt, va_list ap)
 ///  - memformat(15300)         -> "14.9 KB"
 ///  - memformat(15300000)      -> "14.6 MB"
 ///  - memformat(15300000000LL) -> "14.2 GB"
-std::string OIIO_API memformat (long long bytes, int digits=1);
+std::string OIIO_API memformat (uint64_t bytes, int digits=1);
 
 /// Return a string expressing an elapsed time, in human readable form.
 /// e.g. "0:35.2"
